Add parser for user native
