+++ src/native/mod.rs
@@ -2,3 +2,4 @@ mod common;
 mod deserializers;
 pub mod site_native;
 pub mod subject_native;
+pub mod user_native;
