+++ src/native/user_native.rs
@@ -1,0 +1,68 @@
+use serde::Deserialize;
+
+#[cfg(feature = "python")]
+use pyo3::prelude::*;
+
+pub use crate::native::{
+    common::{Category, Entry, Field, Form, Reason, State, Value},
+    deserializers::{
+        default_datetime_none, default_string_none, deserialize_empty_string_as_none,
+        deserialize_empty_string_as_none_datetime,
+    },
+};
+
+#[cfg(not(feature = "python"))]
+#[derive(Debug, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct User {
+    pub unique_id: String,
+
+    #[serde(
+        default = "default_string_none",
+        deserialize_with = "deserialize_empty_string_as_none"
+    )]
+    pub last_language: Option<String>,
+    pub creator: String,
+    pub number_of_forms: usize,
+
+    #[serde(rename = "form")]
+    pub forms: Option<Vec<Form>>,
+}
+
+#[cfg(feature = "python")]
+#[derive(Clone, Debug, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+#[pyclass(get_all)]
+pub struct User {
+    pub unique_id: String,
+
+    #[serde(
+        default = "default_string_none",
+        deserialize_with = "deserialize_empty_string_as_none"
+    )]
+    pub last_language: Option<String>,
+    pub creator: String,
+    pub number_of_forms: usize,
+
+    #[serde(rename = "form")]
+    pub forms: Option<Vec<Form>>,
+}
+
+#[cfg(not(feature = "python"))]
+/// Contains the information from the Prelude native user XML.
+#[derive(Debug, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+pub struct UserNative {
+    #[serde(rename = "user", default)]
+    pub users: Vec<User>,
+}
+
+#[cfg(feature = "python")]
+/// Contains the information from the Prelude native user XML.
+#[derive(Debug, Deserialize, PartialEq)]
+#[serde(rename_all = "camelCase")]
+#[pyclass(get_all)]
+pub struct UserNative {
+    #[serde(rename = "user", default)]
+    pub users: Vec<User>,
+}
