+++ tests/assets/user_native.xml
@@ -1,0 +1,58 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<export_from_vision_EDC date="02-Jun-2024 06:59 -0500" createdBy="Paul Sanders" role="Project Manager" numberSubjectsProcessed="3">
+
+  <user uniqueId="1691421275437" lastLanguage="" creator="Paul Sanders(1681162687395)" numberOfForms="1">
+    <form name="form.name.demographics" lastModified="2023-08-07 10:15:41 -0500" whoLastModifiedName="Paul Sanders" whoLastModifiedRole="Project Manager" whenCreated="1691421341578" hasErrors="false" hasWarnings="false" locked="false" user="" dateTimeChanged="" formTitle="User Demographics" formIndex="1" formGroup="" formState="In-Work">
+      <state value="form.state.in.work" signer="Paul Sanders - Project Manager" signerUniqueId="1681162687395" dateSigned="2023-08-07 10:15:41 -0500" />
+      <category name="demographics" type="normal" highestIndex="0">
+        <field name="address" type="text" dataType="string" errorCode="undefined" whenCreated="2024-01-12 14:14:09 -0600" keepHistory="true" />
+        <field name="email" type="text" dataType="string" errorCode="undefined" whenCreated="2023-08-07 10:15:41 -0500" keepHistory="true">
+          <entry id="1">
+            <value by="Paul Sanders" byUniqueId="1681162687395" role="Project Manager" when="2023-08-07 10:15:41 -0500" xml:space="preserve">jazz@artemis.com</value>
+          </entry>
+        </field>
+      </category>
+      <category name="Administrative" type="normal" highestIndex="0">
+        <field name="study_assignment" type="text" dataType="" errorCode="undefined" whenCreated="2023-08-07 10:15:41 -0500" keepHistory="true">
+          <entry id="1">
+            <value by="set from calculation" byUniqueId="" role="System" when="2023-08-07 10:15:41 -0500" xml:space="preserve">On 07-Aug-2023 10:15 -0500, Paul Sanders assigned user from another study</value>
+            <reason by="set from calculation" byUniqueId="" role="System" when="2023-08-07 10:15:41 -0500" xml:space="preserve">calculated value</reason>
+          </entry>
+        </field>
+      </category>
+    </form>
+  </user>
+
+  <user uniqueId="1681162687395" lastLanguage="" creator="1609858291483(1609858291483)" numberOfForms="2">
+    <form name="form.name.demographics" lastModified="2023-04-10 16:39:30 -0500" whoLastModifiedName="1609858291483" whoLastModifiedRole="role.administrator" whenCreated="1681162770378" hasErrors="false" hasWarnings="false" locked="false" user="" dateTimeChanged="" formTitle="User Demographics" formIndex="1" formGroup="" formState="In-Work">
+      <state value="form.state.in.work" signer="1609858291483 - Administrator" signerUniqueId="1609858291483" dateSigned="2023-04-10 16:39:30 -0500" />
+      <category name="demographics" type="normal" highestIndex="0">
+        <field name="address" type="text" dataType="string" errorCode="undefined" whenCreated="2023-04-15 10:21:14 -0500" keepHistory="true" />
+        <field name="email" type="text" dataType="string" errorCode="undefined" whenCreated="2023-04-10 16:39:30 -0500" keepHistory="true">
+          <entry id="1">
+            <value by="1609858291483" byUniqueId="1609858291483" role="Administrator" when="2023-04-10 16:39:30 -0500" xml:space="preserve">paul@pbsdatasolutions.com</value>
+          </entry>
+        </field>
+      </category>
+      <category name="Administrative" type="normal" highestIndex="0">
+        <field name="study_assignment" type="text" dataType="" errorCode="undefined" whenCreated="2023-04-10 16:39:30 -0500" keepHistory="true">
+          <entry id="1">
+            <value by="set from calculation" byUniqueId="" role="System" when="2023-04-10 16:39:30 -0500" xml:space="preserve">On 10-Apr-2023 16:39 -0500, 1609858291483 assigned user from another study</value>
+            <reason by="set from calculation" byUniqueId="" role="System" when="2023-04-10 16:39:30 -0500" xml:space="preserve">calculated value</reason>
+          </entry>
+        </field>
+      </category>
+    </form>
+    <form name="training.form.name.checklist" lastModified="" whenCreated="1702998014133" hasErrors="false" hasWarnings="false" locked="false" user="" dateTimeChanged="" formTitle="Training Checklist" formIndex="1" formGroup="Training" formState="New">
+      <state value="form.state.new" signer="System set" signerUniqueId="system" dateSigned="2023-12-19 09:00:14 -0600" />
+      <category name="Training_Modules" type="normal" highestIndex="0">
+        <field name="date(1)" type="setToday" dataType="date" errorCode="undefined" whenCreated="2023-12-19 09:00:15 -0600" keepHistory="true" />
+      </category>
+      <category name="Other_Training_Modules" type="indexed" highestIndex="1">
+        <field name="date_field(1)" type="setToday" dataType="date" errorCode="undefined" whenCreated="2023-12-19 09:00:15 -0600" keepHistory="true" />
+      </category>
+      <category name="21CFR11" type="normal" highestIndex="0" />
+    </form>
+  </user>
+
+</export_from_vision_EDC>
